Added defaultPath and defaultApiVersion to the client's constructor
The node-onfleet wrapper defaults the DEFAULT_PATH and DEFAULT_API_VERSION, but may we need other values in some cases.

In order to enable these, I'll add two params to the client's constructor that will default to these constants. If we don't pass anything for these params everything will work as usual.